services/usermgmt/requirements.txt
flask==3.0.3
flask-cors==4.0.1
pymongo==4.8.0
python-dotenv==1.0.1
certifi==2024.2.2
bcrypt==4.1.3